requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
openai==1.3.0
assemblyai==0.21.0
runwayml==0.1.0 